backend/requirements.txt
# FastAPI + server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.2.0

# AI clients (lazy-loaded, all optional at startup)
anthropic>=0.28.0
mistralai>=1.0.0
google-generativeai>=0.7.0
Pillow>=10.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0       # local feedback DB (PostgreSQL)
aiomysql>=0.2.0       # AlgoPython source DB (MySQL)

# RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0
numpy>=1.26.0
